Making profile and posts look better, added badges

+++ db/create-db.sql
@@ -23,7 +23,7 @@ CREATE TABLE post
     id SERIAL PRIMARY KEY,
     user_id INTEGER NOT NULL REFERENCES "user",
     price FLOAT NOT NULL,
-    quantity INTEGER NOT NULL,
+    quantity FLOAT NOT NULL,
     unit VARCHAR(5) NOT NULL,
     CHECK (unit = 'item' OR
            unit = 'oz' OR
